+++ target_snowflake/exceptions.py
@@ -23,3 +23,11 @@ class FileFormatNotFoundException(Exception):
 
 class InvalidFileFormatException(Exception):
     """Exception to raise when name file format is not compatible"""
+
+
+class UnexpectedMessageTypeException(Exception):
+    """Exception to raise when provided message doesn't match the expected type"""
+
+
+class PrimaryKeyNotFoundException(Exception):
+    """Exception to raise when primary key not found in the record message"""